3_StoryForge_Agent/pyproject.toml
[project]
name = "3-storyforge-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.19",
]
